+++ packages/rust-bindings/scripts/build-react-native-wallet.ts
@@ -1,0 +1,9 @@
+import path from 'node:path';
+
+import { convertWasmToJs, copyToFolder } from './build-scripts';
+
+const walletBundlerPath = path.join(__dirname, '../lib/wallet/bundler');
+const walletOutPath = path.join(__dirname, '../lib/wallet/react-native');
+
+copyToFolder(walletBundlerPath, walletOutPath);
+convertWasmToJs(walletOutPath);