+++ examples/nodejs/client/getTokenList.ts
@@ -1,0 +1,57 @@
+import { BlockHash } from '@concordium/web-sdk';
+import { ConcordiumGRPCNodeClient } from '@concordium/web-sdk/nodejs';
+import { TokenId } from '@concordium/web-sdk/plt';
+import { credentials } from '@grpc/grpc-js';
+import meow from 'meow';
+
+import { parseEndpoint } from '../shared/util.js';
+
+const cli = meow(
+    `
+  Usage
+    $ yarn run-example <path-to-this-file> [options]
+
+  Options
+    --help,         Displays this message
+    --block,    -b  A block to query from, defaults to last final block
+    --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
+`,
+    {
+        importMeta: import.meta,
+        flags: {
+            block: {
+                type: 'string',
+                alias: 'b',
+            },
+            endpoint: {
+                type: 'string',
+                alias: 'e',
+                default: 'localhost:20000',
+            },
+        },
+    }
+);
+
+const [address, port] = parseEndpoint(cli.flags.endpoint);
+
+const client = new ConcordiumGRPCNodeClient(address, Number(port), credentials.createInsecure());
+
+/**
+ * Retrieves the protocol level tokens that exists at the end of a given block as an async
+ * iterable. If a blockhash is not supplied it will pick the latest finalized
+ * block. An optional abortSignal can also be provided that closes the stream.
+
+ * Note: A stream can be collected to a list with the streamToList function.
+ */
+
+(async () => {
+    // #region documentation-snippet
+    const blockHash = cli.flags.block === undefined ? undefined : BlockHash.fromHexString(cli.flags.block);
+    const tokens: AsyncIterable<TokenId.Type> = client.getTokenList(blockHash);
+    // #endregion documentation-snippet
+
+    console.log('Protocol level tokens (PLTs) that exists at the end of the given block:');
+    for await (const token of tokens) {
+        console.log(token.toString());
+    }
+})();