+++ packages/sdk/src/constants.ts
@@ -1,3 +1,5 @@
 import * as Energy from './types/Energy.js';
 
 export const DEFAULT_INVOKE_ENERGY: Energy.Type = Energy.create(1000000n);
+export const MAX_U64 = 18446744073709551615n; // 2n ** 64n - 1n
+export const MAX_U32 = 4294967295; // 2n ** 32n - 1n