+++ packages/sdk/src/web3-id/proofs.ts
@@ -1,6 +1,7 @@
 import { Buffer } from 'buffer/index.js';
 
 import { EU_MEMBERS, MAX_DATE, MIN_DATE, StatementBuilder, StatementTypes } from '../commonProofTypes.js';
+import { MAX_U64 } from '../constants.js';
 import { getPastDate } from '../id/idProofs.js';
 import { AttributeKey, AttributeKeyString, AttributeList, AttributesKeys, HexString, Network } from '../types.js';
 import type * as ContractAddress from '../types/ContractAddress.js';
@@ -33,7 +34,6 @@ import {
 } from './types.js';
 
 export const MAX_STRING_BYTE_LENGTH = 31;
-export const MAX_U64 = 18446744073709551615n; // 2n ** 64n - 1n
 export const MIN_DATE_ISO = '-262144-01-01T00:00:00Z';
 export const MAX_DATE_ISO = '+262143-12-31T23:59:59.999999999Z';
 export const MIN_DATE_TIMESTAMP = Date.parse(MIN_DATE_ISO);