+++ packages/sdk/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- Removed build entrypoint split for react-native
-- Removed wasm dir as separate path in package
+### Added
+- The "wasm" entrypoint `@concordium/web-sdk/wasm` from now supports react-native. This means that the partial support for react-native is now extended to full support.
 
 ## 9.0.0